remove lazy modules and move to experimental
