+++ oslo/torch/nn/__init__.py
@@ -1,4 +1,4 @@
-from oslo.torch.nn.modules.conv import Conv1D, LazyConv1D
+from oslo.torch.nn.modules.conv import Conv1D
 from oslo.torch.nn.modules.dropout import (
     FusedBiasDropout,
     FusedBiasDropoutResidual,
@@ -6,7 +6,6 @@ from oslo.torch.nn.modules.dropout import (
 from oslo.torch.nn.modules.embedding import (
     Embedding2D,
     Embedding2p5D,
-    LazyEmbedding,
     VocabParallelEmbedding1D,
     VocabParallelEmbedding2D,
     VocabParallelEmbedding2p5D,
@@ -28,7 +27,6 @@ from oslo.torch.nn.modules.functional import (
 from oslo.torch.nn.modules.layer_norm import LayerNorm2D, LayerNorm2p5D
 from oslo.torch.nn.modules.linear import (
     ColumnParallelLinear,
-    LazyLinear,
     Linear,
     Linear2D,
     Linear2p5D,
